# adidas-subscription-api/src/main/resources/api.yaml
openapi: 3.0.0
info:
  description: API for adidas-subscription-api
  version: "1.0.0-oas3"
  title: Subscription API 
tags:
  - name: Subscription api
    description: subscription api for adidas
paths:
  /subscription:
    post:
      tags:
        - Subscription api
      summary: >-
        creaate a new subscription for given informations as email, firstname, geneder, dateofbirth, consent, newsletterId.
      operationId: newSubscription
      requestBody:
        description: request body required for creation of new subscriptions
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
    get:
      tags:
        - Subscription api
      summary: >-
        get a subscription over given email
      operationId: getSubscription
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
    delete:
      tags:
        - Subscription api
      summary: >-
        delete a subscription over given email
      operationId: deleteByEmail
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
                
  /subscription/all:  
    get:
      tags:
        - Subscription api
      summary: >-
        get all subscriptions
      operationId: getAllSubscription
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionDetailedResponse' 
    delete:
      tags:
        - Subscription api
      summary: >-
        delete all subscription 
      operationId: deleteAll
      responses:
        '200':
          description: Success            

components:
  schemas:
      SubscriptionDetailedResponse:
        properties:
          count:
            type: integer
          subscriptionsList:
            type: array
            items:
              type: object
              $ref: '#/components/schemas/SubscriptionDto'
      SubscriptionResponse:
        properties:
          subscriptionId:
            type: integer
            format: int64
          emailSent:
            type: boolean  
          subscriptionDto:
            type: object
            $ref: '#/components/schemas/SubscriptionDto'
      SubscriptionRequest:
        properties:
          subscriptionDto:
            type: object
            $ref: '#/components/schemas/SubscriptionDto'
      SubscriptionDto:
        properties:
          email:
            type: string
          firstName:
            type: string
          gender:
            type: string
          dateOfBirth:
            type: string
          consent:
            type: boolean
          newsletterId:
            type: integer
      ResponseError:
        properties:
          errorCode:
            type: integer
          errorMessage:
            type: string       
          detailedMessage:
            type: string
          exceptionName:
            type: string
          createdAt:
            type: string
      EmailResponse:
        properties:
          message: 
            type: string
          timestamp:
            type: string
          success: 
            type: boolean
      EmailRequest:
        properties:
          senderEmail:
            type: string
          senderEmailPassword:
            type: string  
          recieverEmail:
            type: string
          senderName:
            type: string  
          subject:
            type: string
          message:
            type: string    
          newsletterId:
            type: integer
            format: int64          